requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
plotly>=5.19.0
openpyxl>=3.1.2
anthropic>=0.21.3
openai>=1.30.0
python-dotenv>=1.0.1
faker>=24.0.0
xlsxwriter>=3.1.9
pyarrow>=14.0.0
